refactor: Relocate navigation-related files to common/navigation (#49)
