+++ app/src/main/java/com/example/rentit/MainActivity.kt
@@ -11,7 +11,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
-import com.example.rentit.common.component.NavigationRoutes
+import com.example.rentit.common.navigation.NavigationRoutes
 import com.example.rentit.common.storage.getToken
 import com.example.rentit.common.theme.RentItTheme
 import com.example.rentit.presentation.MainView
